:recycle: Move auth logic to shared
